Add GetIstioMeshConfig method into IstioClientInterface
