+++ kubernetes/kubetest/mock.go
@@ -436,6 +436,11 @@ func (o *K8SClientMock) GetAuthorizationDetails(namespace string) (*kubernetes.R
 	return args.Get(0).(*kubernetes.RBACDetails), args.Error(1)
 }
 
+func (o *K8SClientMock) GetIstioConfigMap() (*kubernetes.IstioMeshConfig, error) {
+	args := o.Called()
+	return args.Get(0).(*kubernetes.IstioMeshConfig), args.Error(1)
+}
+
 func (o *K8SClientMock) IsOpenShift() bool {
 	args := o.Called()
 	return args.Get(0).(bool)
